requirements.txt
aiogram==2.25.1
selenium
aiosqlite
python-dotenv
aiohttp
parse